modification update User Records

+++ src/hooks/useProfile.ts
@@ -5,8 +5,10 @@ import { v4 as uuidv4 } from 'uuid'
 import { updateProfile } from 'firebase/auth'
 import { useAuth } from '../context/AuthContext'
 import { deleteFile } from '../firebase/deleteFile'
-// import { uploadFile } from '../firebase/uploadFile'
-import { updateUserRecords } from '../firebase/updateUserRecords'
+import { uploadFile } from '../firebase/uploadFile'
+import updateUserRecords from '../firebase/updateUserRecords'
+import { TImagesObj, TUserObj } from '../types/profileType'
+// import uploadFile from '../firebase/uploadFile'
 
 export const useProfile = () => {
   const { currentUsers, setIsLoading } = useAuth()
@@ -15,6 +17,7 @@ export const useProfile = () => {
   const [fileState, setFileState] = useState<any | null>(null)
   const [photoURL, setPhotoURL] = useState<any>(currentUsers?.photoURL)
   const navigate = useNavigate()
+
   const handleChange = (e: any) => {
     const photo = e.target.files[0]
     if (photo) {
@@ -22,34 +25,44 @@ export const useProfile = () => {
       setPhotoURL(URL.createObjectURL(photo))
     }
   }
-  const handleSubmit = async () => {
+
+  const handleSubmit = async (e: any) => {
+    e.preventDefault()
     setIsLoading(true)
 
-    let userObj = { displayName: name, isAdmin: false }
-    let imagesObj = { uName: name }
+    let userObj: TUserObj = {
+      displayName: name,
+      isAdmin: false,
+      photoURL: '',
+    }
+    let imagesObj: TImagesObj = { uName: name }
+
     try {
       if (fileState) {
+        // add photo
         const imageName = uuidv4() + fileState?.name.split('.').pop()
-        // const url = await uploadFile(
-        //   fileState,
-        //   `profile/${currentUsers?.uid}/${imageName}`
-        // )
-        console.log(imageName)
+        const url = await uploadFile(
+          fileState,
+          `profile/${currentUsers?.uid}/${imageName}`
+        )
+        // remove photo
         if (currentUsers?.photoURL) {
-          const prevImage = currentUsers?.photoURL
-            ?.split(`${currentUsers?.uid}%2F`)[1]
+          const prevImage = decodeURIComponent(currentUsers.photoURL)
+            .split(`${currentUsers.uid}/`)[1]
             .split('?')[0]
-          try {
-            await deleteFile(`profile/${currentUsers?.uid}/${prevImage}`)
-          } catch (error) {
-            toast.error('something went wrong')
+          if (prevImage) {
+            try {
+              await deleteFile(`profile/${currentUsers?.uid}/${prevImage}`)
+            } catch (error) {
+              console.log(error)
+            }
           }
         }
-        const url =
-          'https://firebasestorage.googleapis.com/v0/b/react-http-c10cb.appspot.com/o/profile%2FMxnctT4zNdQnzWHYwTjCApHS1AJ3%2F44547b2c-6f70-4488-9ce4-15aad0174521jfif?alt=media&token=83505e47-2944-413e-a750-31a40a2659c0'
-        userObj = { photoURL: url, isAdmin: true } as any
-        imagesObj = { ...imagesObj, uPhoto: url } as any
+        userObj = { ...userObj, photoURL: url, isAdmin: false }
+        imagesObj = { ...imagesObj, uPhoto: url }
       }
+
+      // aktualizacja imienia oraz photo
       if (currentUsers) {
         await updateProfile(currentUsers, userObj)
         await updateUserRecords('Users', currentUsers?.uid, imagesObj)
